Restore SIMD support
The previous commit tested a change to the logic with only two of the
previously supported SIMD flavors. Now that it works, expand the
approach to treat the full range again.

Also stopped downloaded the regressiontests, because the tests are not
run.

A hook script that used GROMACS_ARCH was recently added. That's now
broken. The purpose of that script is unclear, but could be fixable.

Fixes #4